requirements.txt
numpy==1.24.3
opencv_python==4.7.0.72
paho_mqtt==1.6.1
PyYAML==6.0
tflite_support==0.4.4
requests==2.30.0
Pillow==9.5.0